refactor: change project
- 마이페이지 부분 다시 나눔
- 여기서 무조건 재사용성 가능한 컴포넌트로 대처함

obab-client #8
