+++ src/component/my-page/web/item/MyPageRightComponent.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Outlet} from "react-router-dom";
-import MyPageContainer from "../../../../../hooks/MyPageContainer";
+import MyPageContainer from "../../../../hooks/MyPageContainer";
 
 const MyPageRightComponent = () => {
   const {onClickMyPageMenuBar} = MyPageContainer();
