+++ src/page/my-page/WriteBoardPage.jsx
@@ -13,8 +13,7 @@ const CardMediaStyled = styled(CardMedia)({
   paddingTop: '56.25%', // 16:9 비율
 });
 
-const WriteBoardWebComponent = () => {
-  let arr = ['', '', '', '', '', '']
+const WriteBoardPage = () => {
   return (
     <Container maxWidth="lg" className={`mt-8`}>
       <Grid container spacing={4} className="mt-8">
@@ -44,28 +43,7 @@ const WriteBoardWebComponent = () => {
           </Grid>
         </Grid>
       </Grid>
-    </Container>
-  );
-//   return (
-//     <div className={`${styles?.myPageRouterComponent}`}>
-//       <div className={`${styles?.writeBoardContents} flex flex-wrap`}>
-//         {
-//           arr.map(() =>
-//             <div className={`flex flex-col h-96 w-60 border-4 mx-5 my-5`}>
-//               <img src={`/images/siteinfo/siteinfo.webp`} alt={`이미지`} className={`h-52 border-4`}/>
-//               <div className={`text-left text-xl py-1 ${styles?.commonEllipsis}`}>{`Easy and Quick Recipes with
-// Ingredients From Your Fridge`}</div>
-//               <div className={`text-left py-1 text-sm text-gray-400 ${styles?.commonEllipsis}`}>{`March 5, 2024`}</div>
-//               <div className={`text-left py-1 text-gray-500 ${styles?.commonEllipsis}`}>{`OBAB is a Korean food and recipe blog that
-// offers a unique…`}</div>
-//             </div>
-//           )
-//         }
-//       </div>
-//
-//       <div className={`m-10 text-blue-400 cursor-pointer text-lg`}>{`더 보기`}</div>
-//     </div>
-//   );
+    </Container>  )
 }
 
-export default WriteBoardWebComponent;
+export default WriteBoardPage;
