--- src/page/my/page/InterestedBoardPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import InterestedBoardWebComponent from "../../../component/my/page/interested/board/InterestedBoardWebComponent";
-
-const InterestedBoardPage = () => {
-  return (
-    <InterestedBoardWebComponent/>
-  )
-}
-
-export default InterestedBoardPage;
