--- src/page/my/page/WriteBoardPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import WriteBoardWebComponent from "../../../component/my/page/write/board/WriteBoardWebComponent";
-
-const WriteBoardPage = () => {
-  return (
-    <WriteBoardWebComponent/>
-  )
-}
-
-export default WriteBoardPage;
