--- src/page/my/page/WriteCommentPage.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import WriteCommentWebComponent from "../../../component/my/page/write/comment/WriteCommentWebComponent";
-
-const WriteCommentPage = () => {
-  return (
-    <WriteCommentWebComponent/>
-  )
-}
-
-export default WriteCommentPage;
