+++ src/router/router.jsx
@@ -5,9 +5,9 @@ import CategoryViewPage from "../page/recipe/RecipeViewPage";
 import ErrorPage from "../page/error/ErrorPage";
 import Login from "../page/login/user/Login";
 import MyPageLayout from "../layout/my/page/MyPageLayout";
-import InterestedBoardPage from "../page/my/page/InterestedBoardPage";
-import WriteBoardPage from "../page/my/page/WriteBoardPage";
-import WriteCommentPage from "../page/my/page/WriteCommentPage";
+import InterestedBoardPage from "../page/my-page/InterestedBoardPage";
+import WriteBoardPage from "../page/my-page/WriteBoardPage";
+import WriteCommentPage from "../page/my-page/WriteCommentPage";
 import MainLayout from "../layout/main/MainLayout";
 import RecipeListPage from "../page/recipe/RecipeListPage";
 import SignUp from "../page/login/user/SignUp";
